+++ seans_clothing/public/index.html
@@ -24,6 +24,12 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Bigelow+Rules&family=Nunito:wght@200&family=Open+Sans:wght@300&family=Permanent+Marker&family=Roboto:wght@300&family=Ubuntu:wght@400;700&display=swap"
+      rel="stylesheet"
+    />
     <title>React App</title>
   </head>
   <body>
